fix(inference): address multiple issues causing runtime failures

+++ src/backend/ops/embedding/cpu/embedding_cpu.cpp
@@ -7,7 +7,7 @@
 #include <cstring>
 
 template <typename T>
-void embedding_(T *out, const int64_t *index, const T *weight, size_t numel, size_t len) {
+void embedding_(T *out, const int *index, const T *weight, size_t numel, size_t len) {
     const size_t nlen = numel / len;
 
     // Avoid OpenMP for memory-bound operations;
@@ -22,12 +22,12 @@ namespace neollm::ops::cpu {
 void embedding(std::byte *out, const std::byte *index, const std::byte *weight, NeollmDataType_t type, size_t numel, size_t len) {
     switch (type) {
     case NEOLLM_DTYPE_F32:
-        return embedding_(reinterpret_cast<float *>(out), reinterpret_cast<const int64_t *>(index), reinterpret_cast<const float *>(weight), numel, len);
+        return embedding_(reinterpret_cast<float *>(out), reinterpret_cast<const int *>(index), reinterpret_cast<const float *>(weight), numel, len);
     case NEOLLM_DTYPE_BF16:
-        return embedding_(reinterpret_cast<neollm::bf16_t *>(out), reinterpret_cast<const int64_t *>(index),
+        return embedding_(reinterpret_cast<neollm::bf16_t *>(out), reinterpret_cast<const int *>(index),
                           reinterpret_cast<const neollm::bf16_t *>(weight), numel, len);
     case NEOLLM_DTYPE_F16:
-        return embedding_(reinterpret_cast<neollm::fp16_t *>(out), reinterpret_cast<const int64_t *>(index),
+        return embedding_(reinterpret_cast<neollm::fp16_t *>(out), reinterpret_cast<const int *>(index),
                           reinterpret_cast<const neollm::fp16_t *>(weight), numel, len);
     default:
         EXCEPTION_UNSUPPORTED_DATATYPE(type);
